requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
EbookLib==0.18
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
macholib==1.16.2
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.0
Pygments==2.14.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
python-dotenv==1.0.0
requests==2.28.2
rich==13.3.2
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
